+++ emu/app/Tree.scala
@@ -1,6 +1,6 @@
 sealed abstract class Tree
 object Tree {
-  case class Num(n: Int) extends Tree
+  case class Num(n: Long) extends Tree
 
   case class F1(name: String) extends Tree
   case class F2(name: String) extends Tree
